+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [Unreleased]
+
+### Changed
+- fix(daemon-monitor): detect_active_stories emits 3phase stories post Option-A
+
 ## [2.23.0] - 2026-05-06
 
 ### Changed